Move MockSession into a separate file
